fix(seo): bound Gemma reasoning budget
Gemma 4 reasoning could consume the SEO completion allowance before emitting canonical metadata.
The SEO client now caps thinking while leaving ample output budget.

- Capture the wire request and keep focused reliability tests within the repository LOC limit.
